moved all exceptions to exceptions namespace
