+++ Bot_NetCore/Exceptions/InvalidXMLException.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace Bot_NetCore.Exceptions
+{
+    public class InvalidXMLException : Exception
+    {
+        public InvalidXMLException() : base()
+        {
+        }
+
+        public InvalidXMLException(string message) : base(message)
+        {
+        }
+
+        public InvalidXMLException(string message, Exception inner) : base(message, inner)
+        {
+        }
+    }
+}
