+++ Bot_NetCore/Exceptions/MemberNotFoundException.cs
@@ -1,0 +1,19 @@
+﻿using System;
+
+namespace Bot_NetCore.Exceptions
+{
+    public class MemberNotFoundException : Exception
+    {
+        public MemberNotFoundException() : base()
+        {
+        }
+
+        public MemberNotFoundException(string message) : base(message)
+        {
+        }
+
+        public MemberNotFoundException(string message, Exception inner) : base(message, inner)
+        {
+        }
+    }
+}
